Change structure of option process and some idea.
Remove the idea of let user specify the version of the application to prevent I/O error or misdeterment the version of executable. In the new idea, there can be spilt to two parts, one is without any argument, meaning user only needs to execute the script and it will find the application from the applications folder under root or applications folder under current user home directory. In the second part is let user add "a" flag to patch specific Sketch.app under any directory. Both parts are using hash value to verify the application, one of reason to using this method is to determent the version of Sketch, the second reason is to prevent script mispatch or repatch. Finally, to successfully patch, the executable has to be clean meaning without any modification and directly downloaded from the official website.