Remove the hex prefix chars in args
Remove all `0x` prefix in all args, because the prefix does not help on 
dev processing, and since the address does not match the address in 
Ghidra.app, so there is no point let `0x` prefix hardcode in script 
anymore. But we will let prefix be specified in the `patch` and 
`patchLegacy` processes.